\name{CommonMean.Copula-package}
\alias{CommonMean.Copula-package}
\docType{package}
\title{
\packageTitle{CommonMean.Copula}
}
\description{
Estimate bivariate common mean vector under copula models with known correlation. A maximum likelihood estimation procedure is employed. In the current version, available copulas are the Clayton, Farlie-Gumbel-Morgenstern (FGM), and normal copulas. See Shih et al. (2019) and Shih et al. (2021) for details under the FGM and general copulas, respectively.
}
\details{
The method implemented in this package can be used for bivariate meta-analyses. See Shih et al. (2019) and Shih et al. (2021) for the example of bivariate entrance exam data analysis.
}
\author{
\packageAuthor{CommonMean.Copula}

Maintainer: \packageMaintainer{CommonMean.Copula}
}
\references{
Shih J-H, Konno Y, Chang Y-T, Emura T (2019) Estimation of a common mean vector in bivariate meta-analysis under the FGM copula, Statistics 53(3): 673-95.

Shih J-H, Konno Y, Emura T (2021-) Copula-based estimation methods for a common mean vector for bivariate meta-analyses, under review.
}
