\name{CommonMean.Copula-package}
\alias{CommonMean.Copula-package}
\docType{package}
\title{
\packageTitle{CommonMean.Copula}
}
\description{
Estimate bivariate common mean vector under copula models with known correlation. A maximum likelihood estimation procedure is employed. In the current version, available copulas include the Clayton, Farlie-Gumbel-Morgenstern (FGM), and Gaussian copulas. See Shih et al. (2019) for details under the FGM copula.
}
\details{
The method implemented in this package can be used for bivariate meta-analysis. See Shih et al. (2019) for an example of bivariate entrance exam data analysis under the FGM copula.
}
\author{
\packageAuthor{CommonMean.Copula}

Maintainer: \packageMaintainer{CommonMean.Copula}
}
\references{
Shih J-H, Konno Y, Chang Y-T, Emura T (2019) Estimation of a common mean vector in bivariate meta-analysis under the FGM copula, Statistics 53(3): 673-95.
}

